backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
coverage==7.7.0
decorator==5.2.1
distlib==0.3.9
Django==4.2.17
django-cors-headers==4.6.0
django-extensions==3.2.3
django-redis==5.4.0
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework-stubs==3.15.3
executing==2.2.0
filelock==3.18.0
gunicorn==23.0.0
identify==2.6.9
idna==3.10
iniconfig==2.0.0
ipdb==0.13.13
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pipenv==2024.4.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.2.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
setuptools==76.1.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.29.3
wcwidth==0.2.13
